Fix compilation issue

+++ 0x1E-search_algorithms/search_algos.h
@@ -42,7 +42,7 @@ typedef struct skiplist_s
 } skiplist_t;
 
 int linear_search(int *array, size_t size, int value);
-void print_array(int *array, int l, int r);
+void print_array(int *array, size_t l, size_t r);
 int binary_search_index(int *array, size_t size, int value, int offset);
 int binary_search_index1(int *array, size_t l, size_t r, int value);
 int binary_search(int *array, size_t size, int value);
